refactor: CustomOverlay 생성 로직을 createCustomOverlay로 개선
